samples/jax/requirements.txt
absl-py==0.13.0
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
chex==0.0.8
click==8.0.1
cloudpickle==1.6.0
configparser==5.0.2
cycler==0.10.0
dm-tree==0.1.6
docker-pycreds==0.4.0
filelock==3.0.12
flatbuffers==1.12
flax==0.3.4
future==0.18.2
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.24
glcontext==2.3.4
glfw==1.12.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.0
gym==0.18.0
gym3==0.3.3
h5py==3.1.0
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.3.0
jax==0.2.17
jaxlib==0.1.65+cuda110
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
moderngl==5.6.4
msgpack==1.0.2
numpy==1.21.0
oauthlib==3.1.1
opt-einsum==3.3.0
optax==0.0.9
pathtools==0.1.2
Pillow==8.3.2
procgen==0.10.4+416821e.computecanada
promise==2.3
protobuf==3.18.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
sentry-sdk==1.3.1
shortuuid==1.0.1
six==1.15.0
smmap==4.0.0
subprocess32==3.5.4
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.2
tensorflow-estimator==2.5.0
termcolor==1.1.0
toolz==0.11.1
tqdm==4.62.2
typing-extensions==3.7.4.3
urllib3==1.26.6
wandb==0.12.2
Werkzeug==2.0.1
wrapt==1.12.1
yaspin==2.1.0
tensorflow_probability[jax]